build 20220115

+++ src/libxpeccy/cpu/1801vm1/1801vm1.h
@@ -38,7 +38,7 @@ void pdp11_reset(CPU*);
 int pdp11_exec(CPU*);
 // int pdp11_int(CPU*);
 
-xMnem pdp11_mnem(CPU*, unsigned short, cbdmr, void*);
+xMnem pdp11_mnem(CPU*, int, cbdmr, void*);
 xAsmScan pdp11_asm(const char*, char*);
 
 void pdp11_get_regs(CPU*, xRegBunch*);
