+++ src/libxpeccy/cpu/Z80/z80.h
@@ -38,7 +38,7 @@ enum {
 void z80_reset(CPU*);
 int z80_exec(CPU*);
 xAsmScan z80_asm(const char*, char*);
-xMnem z80_mnem(CPU*, unsigned short, cbdmr, void*);
+xMnem z80_mnem(CPU*, int, cbdmr, void*);
 
 void z80_get_regs(CPU*, xRegBunch*);
 void z80_set_regs(CPU*, xRegBunch);
