+++ src/libxpeccy/cpu/i80286/i80286.h
@@ -86,6 +86,6 @@ void i286_mwr(CPU*, xSegPtr, int, unsigned short, int);
 void i286_reset(CPU*);
 int i286_exec(CPU*);
 xAsmScan i286_asm(const char*, char*);
-xMnem i286_mnem(CPU*, unsigned short, cbdmr, void*);
+xMnem i286_mnem(CPU*, int, cbdmr, void*);
 void i286_get_regs(CPU*, xRegBunch*);
 void i286_set_regs(CPU*, xRegBunch);
